+++ nominal/experimental/dataset_utils/__init__.py
@@ -1,0 +1,5 @@
+from nominal.experimental.dataset_utils._dataset_utils import create_dataset_with_uuid
+
+__all__ = [
+    "create_dataset_with_uuid",
+]